+++ yandex-framework/test/sweeper_helpers.go
@@ -3,20 +3,24 @@ package test
 import (
 	"context"
 	"fmt"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/yandex-cloud/go-genproto/yandex/cloud/operation"
-	"github.com/yandex-cloud/terraform-provider-yandex/common"
-	"github.com/yandex-cloud/terraform-provider-yandex/yandex-framework/provider-config"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"log"
 	"os"
 	"strconv"
 	"strings"
+
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
+	"github.com/yandex-cloud/go-genproto/yandex/cloud/operation"
+	"github.com/yandex-cloud/terraform-provider-yandex/common"
+	provider_config "github.com/yandex-cloud/terraform-provider-yandex/yandex-framework/provider-config"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 const (
 	defaultZoneForSweepers = "ru-central1-a"
+	testAccTestsUser       = "yc.terraform.acctest-sa"
+	testResourseNamePrefix = "yc-tf-acc-tests"
 )
 
 type sweeperFunc func(*provider_config.Config, string) error
@@ -115,4 +119,18 @@ func debugLog(format string, v ...interface{}) {
 func isStatusWithCode(err error, code codes.Code) bool {
 	grpcStatus, ok := status.FromError(err)
 	return ok && grpcStatus.Code() == code
+}
+
+func isTestResourseName(name string) bool {
+	return strings.HasPrefix(name, testResourseNamePrefix)
+}
+
+// testResourseName - return randomized name for resourse with common preffix
+func testResourseName(length int) string {
+	saltLen := length - len(testResourseNamePrefix) - 1
+	return fmt.Sprintf(
+		"%s-%s",
+		testResourseNamePrefix,
+		acctest.RandStringFromCharSet(saltLen, acctest.CharSetAlpha),
+	)
 }